[luci] Make dynamic shape determinant function as public
`has_dynamic_shape` in `ConvertNCHWtoNHWCPass` is for determining whether a node has dynamic shape.
Since some of other passes also needs this function, this commit will make it as public.

ONE-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>
